extracts `useFractalNode` loadDAO method to own hook to fix loading issues
