+++ src/hooks/DAO/loaders/useFractalNode.ts
@@ -1,4 +1,4 @@
-import { useLazyQuery, useQuery } from '@apollo/client';
+import { useQuery } from '@apollo/client';
 import { utils } from 'ethers';
 import { useRouter } from 'next/navigation';
 import { useCallback, useEffect, useRef } from 'react';
@@ -9,33 +9,15 @@ import { BASE_ROUTES } from '../../../constants/routes';
 import { logError } from '../../../helpers/errorLogging';
 import { useFractal } from '../../../providers/App/AppProvider';
 import { NodeAction } from '../../../providers/App/node/action';
-import { FractalNode, Node, WithError } from '../../../types';
+import { Node } from '../../../types';
+import { mapChildNodes } from '../../../utils/hierarchy';
 import { useUpdateTimer } from '../../utils/useUpdateTimer';
 import { useLazyDAOName } from '../useDAOName';
 import { useFractalModules } from './useFractalModules';
 
-const mapChildNodes = (_hierarchy: any) => {
-  return _hierarchy.map((node: any) => {
-    return {
-      nodeHierarchy: {
-        parentAddress: node.parentAddress,
-        childNodes: mapChildNodes(node.hierarchy),
-      },
-      daoName: node.name,
-      daoAddress: node.address,
-    };
-  });
-};
-
 const ONE_MINUTE = 60 * 1000;
 
-export const useFractalNode = ({
-  daoAddress,
-  loadOnMount = true,
-}: {
-  daoAddress?: string;
-  loadOnMount?: boolean;
-}) => {
+export const useFractalNode = ({ daoAddress }: { daoAddress?: string }) => {
   // tracks the current valid DAO address; helps prevent unnecessary calls
   const currentValidAddress = useRef<string | undefined>();
   const {
@@ -48,7 +30,6 @@ export const useFractalNode = ({
 
   const lookupModules = useFractalModules();
   const { setMethodOnInterval } = useUpdateTimer(currentValidAddress.current);
-  const [getDAOInfo] = useLazyQuery(DAOQueryDocument);
 
   const invalidateDAO = useCallback(
     (errorMessage: string) => {
@@ -105,39 +86,6 @@ export const useFractalNode = ({
     pollInterval: ONE_MINUTE,
   });
 
-  const loadDao = useCallback(
-    async (_daoAddress: string): Promise<FractalNode | WithError> => {
-      if (utils.isAddress(_daoAddress)) {
-        try {
-          const safe = await safeService.getSafeInfo(_daoAddress);
-          const fractalModules = await lookupModules(safe.modules);
-          const graphNodeInfo = formatDAOQuery(
-            await getDAOInfo({ variables: { daoAddress: _daoAddress } }),
-            safe.address
-          );
-          if (!graphNodeInfo) {
-            logError('graphQL query failed');
-            return { error: 'errorFailedSearch' };
-          }
-          const daoName = await getDaoName(utils.getAddress(safe.address), graphNodeInfo.daoName);
-
-          return Object.assign(graphNodeInfo, {
-            daoName,
-            safe,
-            fractalModules,
-          });
-        } catch (e) {
-          logError(e);
-          return { error: 'errorInvalidSearch' };
-        }
-      } else {
-        // invalid address
-        return { error: 'errorFailedSearch' };
-      }
-    },
-    [safeService, lookupModules, formatDAOQuery, getDAOInfo, getDaoName]
-  );
-
   const updateSafeInfo = useCallback(
     async (_daoAddress: string) => {
       const safeInfo = await safeService.getSafeInfo(utils.getAddress(_daoAddress));
@@ -180,7 +128,7 @@ export const useFractalNode = ({
 
   useEffect(() => {
     const isCurrentAddress = daoAddress === currentValidAddress.current;
-    if (!currentValidAddress.current && loadOnMount) {
+    if (!currentValidAddress.current) {
       if (currentValidAddress.current === undefined) {
         currentValidAddress.current = daoAddress;
       }
@@ -188,7 +136,5 @@ export const useFractalNode = ({
         setDAO(daoAddress);
       }
     }
-  }, [daoAddress, loadOnMount, setDAO, action, currentValidAddress]);
-
-  return { loadDao };
+  }, [daoAddress, setDAO, action, currentValidAddress]);
 };
